Muda las funciones de archivos debido a los modulos
