+++ src/index.js
@@ -1,11 +1,18 @@
-import {activarEventoBoton} from "../src/ui/botonAgregarTarea.js";
+import {activarEventoBoton} from "./ui/botonAgregarTarea.js";
 
-import {agregarNuevaTarea, mostrarErrorTituloTarea, borrarErrorTituloTarea} from "../src/ui/tareas.js";
+import {agregarNuevaTarea, mostrarErrorTituloTarea, borrarErrorTituloTarea, gestionarOpcionesDeLasTareas, gestionarTareaCompleta, gestionarTareaReiniciada, borrarTarea, recrearTareasGuardadas} from "../src/ui/tareas.js";
 
-import {validarTituloNuevaTarea} from "../src/validaciones/validaciones.js";
+import {gestionarCambioDeLista} from "./ui/listasTareas.js";
 
-// Agrega tareas al apretar el botón correspondiente //
+import {validarTituloNuevaTarea} from "./validaciones/validaciones.js";
+
+import {guardarTareasEnLocalStorage, cargarTareasDeLocalStorage, tareasGuardadasEnLocalStorage}  from "../src/storage/to-do-list.js";
+
 const $botonAgregarTarea = document.querySelector("#boton-agregar-tarea");
+const $contenedorTareasPendientes = document.querySelector(".lista-tareas-pendientes");
+const $contenedorTareasCompletas = document.querySelector(".lista-tareas-completas");
+const $opcionesListasDeTareas = document.querySelectorAll(".texto-opcion");
+const $botonVaciarListasTareas = document.querySelector(".opcion-vaciar-listas");
 
 $botonAgregarTarea.addEventListener(("click"), () => {
 	activarEventoBoton($botonAgregarTarea);
@@ -26,10 +33,6 @@ $botonAgregarTarea.addEventListener(("click"), () => {
 	}
 });
 
-// Asigna funcionalidad a las opciones y completado de tareas//
-
-const $contenedorTareasPendientes = document.querySelector(".lista-tareas-pendientes");
-
 $contenedorTareasPendientes.addEventListener(("click"), (e) => {
 	const $tareaSeleccionada = e.target.closest(".tarea");
 
@@ -39,8 +42,6 @@ $contenedorTareasPendientes.addEventListener(("click"), (e) => {
 		gestionarOpcionesDeLasTareas($tareaSeleccionada, e);
 	}
 });
-
-const $contenedorTareasCompletas = document.querySelector(".lista-tareas-completas");
 
 $contenedorTareasCompletas.addEventListener(("click"), (e) => {
 	const $tareaSeleccionada = e.target.closest(".tarea");
@@ -52,179 +53,12 @@ $contenedorTareasCompletas.addEventListener(("click"), (e) => {
 	}
 });
 
-function agregarNivelDePrioridadATarea($tareaSeleccionada, nivelDePrioridad) {
-	if($tareaSeleccionada.classList.contains("prioridad-3")) {
-		$tareaSeleccionada.classList.remove("prioridad-3");
-		$tareaSeleccionada.classList.add(nivelDePrioridad);
-	} else if($tareaSeleccionada.classList.contains("prioridad-2")) {
-		$tareaSeleccionada.classList.remove("prioridad-2");
-		$tareaSeleccionada.classList.add(nivelDePrioridad);
-	} else if($tareaSeleccionada.classList.contains("prioridad-1")) {
-		$tareaSeleccionada.classList.remove("prioridad-1");
-		$tareaSeleccionada.classList.add(nivelDePrioridad);
-	} else{
-		$tareaSeleccionada.classList.add(nivelDePrioridad);
-	}
-}
-
-function cambiarColorNombreTareas($tareaSeleccionada) {
-	const $nombreTarea = $tareaSeleccionada.querySelector(".nombre-tarea");
-
-	if($tareaSeleccionada.classList.contains("prioridad-3")) {
-		$nombreTarea.style.color = "#fff";
-	} else if($tareaSeleccionada.classList.contains("prioridad-2")) {
-		$nombreTarea.style.color = "#fff";
-	} else if($tareaSeleccionada.classList.contains("prioridad-1")) {
-		$nombreTarea.style.color = "#808080";
-	}
-}
-
-function cambiarColorIconoOpcionesTareas($tareaSeleccionada) {
-	const $iconoOpcionesTarea = $tareaSeleccionada.querySelector(".icono-opciones-tareas");
-
-	if($tareaSeleccionada.classList.contains("prioridad-3")) {
-		$iconoOpcionesTarea.style.color = "#fff";
-	} else if($tareaSeleccionada.classList.contains("prioridad-2")) {
-		$iconoOpcionesTarea.style.color = "#fff";
-	} else if($tareaSeleccionada.classList.contains("prioridad-1")) {
-		$iconoOpcionesTarea.style.color = "#808080";
-	}
-}
-
-function borrarTarea($tareaSeleccionada) {
-	$tareaSeleccionada.parentNode.removeChild($tareaSeleccionada);
-}
-
-// Funcionalidad al completar tarea //
-
-function gestionarTareaCompleta($tareaSeleccionada) {
-	const listaALaQueAgregarTarea = "lista-tareas-completas";
-	const nombreTarea = $tareaSeleccionada.querySelector(".nombre-tarea").textContent;
-
-	agregarNuevaTarea(listaALaQueAgregarTarea, nombreTarea);
-	agregarMismasClasesATareaNueva(listaALaQueAgregarTarea, $tareaSeleccionada);
-	borrarTarea($tareaSeleccionada);
-	actualizarDatosTareaGuardada(nombreTarea);
-}
-
-function agregarMismasClasesATareaNueva(contextoLista, $tareaSeleccionada) {
-	const nombreTareaAnterior = ($tareaSeleccionada.querySelector(".nombre-tarea")).textContent;
-	let $tareaNueva;
-
-	const clasesTareaOriginal = $tareaSeleccionada.classList;
-	let listaClasesTareaActualizada = "";
-	
-	clasesTareaOriginal.forEach((clase) => {
-		if(!(clase === "tarea-pendiente" || clase === "tarea-completa")) {
-			listaClasesTareaActualizada += `${clase} `;
-		}
-	});
-
-	if(contextoLista === "lista-tareas-completas"){
-		const $listaTareasCompletas = document.querySelector(".lista-tareas-completas");
-		const $tareasCompletas = $listaTareasCompletas.querySelectorAll(".tarea");
-
-		$tareasCompletas.forEach(($tarea) => {
-			const nombreTareaCompleta = ($tarea.querySelector(".nombre-tarea")).textContent;
-
-			if(nombreTareaAnterior === nombreTareaCompleta){
-				$tareaNueva = $tarea;
-			}
-		});
-
-		listaClasesTareaActualizada += "tarea-completa";
-		$tareaNueva.classList = listaClasesTareaActualizada;
-	} else if(contextoLista === "lista-tareas-pendientes"){
-		const $listaTareasPendientes = document.querySelector(".lista-tareas-pendientes");
-		const $tareasPendientes = $listaTareasPendientes.querySelectorAll(".tarea");
-
-		$tareasPendientes.forEach(($tarea) => {
-			const nombreTareaPendiente = ($tarea.querySelector(".nombre-tarea")).textContent;
-
-			if(nombreTareaAnterior === nombreTareaPendiente){
-				$tareaNueva = $tarea;
-			}
-		});
-
-		listaClasesTareaActualizada += "tarea-pendiente";
-		$tareaNueva.classList = listaClasesTareaActualizada;
-	}
-
-	cambiarColorNombreTareas($tareaNueva);
-	cambiarColorIconoOpcionesTareas($tareaNueva);
-}
-
-// Funcionalidad al reiniciar tarea //
-
-function gestionarTareaReiniciada(e, $tareaSeleccionada){
-	const listaALaQueAgregarTarea = "lista-tareas-pendientes";
-	const nombreTarea = $tareaSeleccionada.querySelector(".nombre-tarea").textContent;
-	
-	agregarNuevaTarea(listaALaQueAgregarTarea, nombreTarea);
-	agregarMismasClasesATareaNueva(listaALaQueAgregarTarea, $tareaSeleccionada);
-	borrarTarea($tareaSeleccionada);
-	actualizarDatosTareaGuardada(nombreTarea);
-}
-
-// Funcionalidad cambio de listas //
-
-const $opcionesListasDeTareas = document.querySelectorAll(".texto-opcion");
-
 $opcionesListasDeTareas.forEach(($opcionListaTarea) => {
 	$opcionListaTarea.addEventListener(("click"), (e) => {
 		const listaSeleccionada = (e.target).textContent;
 		gestionarCambioDeLista(listaSeleccionada);
 	});
 });
-
-function gestionarCambioDeLista (listaSeleccionada) {
-	if(listaSeleccionada === "Pendientes") {
-		const $listaTareasPendientes = document.querySelector(".contenedor-lista-tareas-pendientes");
-		$listaTareasPendientes.id = "";
-
-		const $listaTareasCompletas = document.querySelector(".contenedor-lista-tareas-completas");
-		$listaTareasCompletas.id = "oculto";
-		const $listaTareasTotales = document.querySelector(".contenedor-lista-tareas-totales");
-		$listaTareasTotales.id = "oculto";
-		mostrarListaActiva(listaSeleccionada);
-	} else if(listaSeleccionada === "Completas"){
-		const $listaTareasCompletas = document.querySelector(".contenedor-lista-tareas-completas");
-		$listaTareasCompletas.id = "";
-
-		const $listaTareasPendientes = document.querySelector(".contenedor-lista-tareas-pendientes");
-		$listaTareasPendientes.id = "oculto";
-		const $listaTareasTotales = document.querySelector(".contenedor-lista-tareas-totales");
-		$listaTareasTotales.id = "oculto";
-		mostrarListaActiva(listaSeleccionada);
-	} else if(listaSeleccionada === "Total") {
-		const $listaTareasTotales = document.querySelector(".contenedor-lista-tareas-totales");
-		$listaTareasTotales.id = "";
-
-		const $listaTareasPendientes = document.querySelector(".contenedor-lista-tareas-pendientes");
-		$listaTareasPendientes.id = "oculto";
-		const $listaTareasCompletas = document.querySelector(".contenedor-lista-tareas-completas");
-		$listaTareasCompletas.id = "oculto";
-		mostrarListaActiva(listaSeleccionada);
-	}
-}
-
-function mostrarListaActiva(listaSeleccionada) {
-	const $opcionesListasDeTareas = document.querySelectorAll(".texto-opcion");
-	
-	$opcionesListasDeTareas.forEach(($opcionListaTarea) => {
-		if($opcionListaTarea.classList.contains("opcion-activa")) {
-			$opcionListaTarea.classList.remove("opcion-activa");
-		}
-	});
-
-	$opcionesListasDeTareas.forEach(($opcionListaTarea) => {
-		if($opcionListaTarea.textContent === listaSeleccionada) {
-			$opcionListaTarea.classList.add("opcion-activa");
-		}
-	});
-}
-
-const $botonVaciarListasTareas = document.querySelector(".opcion-vaciar-listas");
 
 $botonVaciarListasTareas.addEventListener(("click"), () => {
 	const $totalDeTareas = document.querySelectorAll(".tarea");
@@ -233,95 +67,6 @@ $botonVaciarListasTareas.addEventListener(("click"), () => {
 		borrarTarea($tarea);
 	});
 });
-
-// Funcionalidad localStorage //
-
-let tareasGuardadasEnLocalStorage = [];
-
-function guardarTareasEnLocalStorage(contextoTarea, nombreTarea){	
-	if(contextoTarea !== "nueva tarea"){
-		localStorage.setItem("tareasGuardadas", JSON.stringify(tareasGuardadasEnLocalStorage));
-	} else{
-		const $tareas = document.querySelectorAll(".tarea");
-
-		$tareas.forEach((tarea) => {
-			let tituloTarea = tarea.querySelector(".nombre-tarea").textContent;
-			const nombreKey = "tarea";
-
-			if(tituloTarea === nombreTarea){
-				const informacionTareaAGuardar = {
-					[nombreKey]: {
-						"nombreTarea": nombreTarea,
-						"clasesTarea": tarea.classList,
-					}
-				};
-				tareasGuardadasEnLocalStorage.push(informacionTareaAGuardar);
-
-				localStorage.setItem("tareasGuardadas", JSON.stringify(tareasGuardadasEnLocalStorage));
-			}
-		});
-	}
-}
-
-function cargarTareasDeLocalStorage(DATA_A_BUSCAR){
-	if(DATA_A_BUSCAR === undefined){
-		throw new Error("Se necesita el nombre de la data que se requiere para cargar la misma");
-	}
-	
-	const tareasGuardadas = JSON.parse(localStorage.getItem(`${DATA_A_BUSCAR}`));
-
-	if(tareasGuardadas === null) {
-		throw new Error(`La data ${DATA_A_BUSCAR} no se encontro en el localStorage`);
-	}
-
-	return tareasGuardadas;
-}
-
-function actualizarDatosTareaGuardada(nombreTareaActualizada){
-	const contextoTarea = "actualización de tarea";
-	let NOMBRE_KEY = "tarea";
-	let tareaActualizada;
-
-	tareasGuardadasEnLocalStorage.forEach((tareaEnLista, i) => {
-		const $tareas = document.querySelectorAll(".tarea");
-		const tituloTareaAComparar = ($tareas[i].querySelector(".nombre-tarea")).textContent;
-
-		if(tituloTareaAComparar === nombreTareaActualizada){
-			tareaActualizada = {
-				[NOMBRE_KEY]: {
-					"nombreTarea": nombreTareaActualizada,
-					"clasesTarea": $tareas[i].classList
-				}
-			};				
-		}
-	});
-
-	tareasGuardadasEnLocalStorage.forEach((tareaEnLista, i) => {
-		const indicadorTareaEnLista = Object.keys(tareaEnLista)[0];
-		const nombreTareaGuardada = tareasGuardadasEnLocalStorage[i][indicadorTareaEnLista]["nombreTarea"];
-
-		if(nombreTareaActualizada === nombreTareaGuardada){
-			tareasGuardadasEnLocalStorage[i] = tareaActualizada;
-		}
-	});
-
-	guardarTareasEnLocalStorage(contextoTarea, nombreTareaActualizada);
-}
-
-function eliminarTareaDeLocalStorage($tareaSeleccionada){
-	const nombreTareaSeleccionada = $tareaSeleccionada.querySelector(".nombre-tarea").textContent;
-
-	tareasGuardadasEnLocalStorage.forEach((tareaGuardada, i) => {
-		const indicadorTareaGuardada = Object.keys(tareaGuardada)[0];
-		const nombreTareaGuardada = tareasGuardadasEnLocalStorage[i][indicadorTareaGuardada]["nombreTarea"];
-
-		if(nombreTareaGuardada === nombreTareaSeleccionada){
-			tareasGuardadasEnLocalStorage.splice(0, i + 1);
-		}
-	});
-}
-
-// Funcionalidad al iniciar la página //
 
 function iniciarPagina() {
 	const DATA_A_BUSCAR = ["tareasGuardadas"];
@@ -346,7 +91,7 @@ function gestionarDataTareasGuardadas(dataTareas){
 	dataTareas.forEach((tareaGuardada) => {
 		const indicadorTarea = Object.keys(tareaGuardada)[0];
 		const listaClasesTareaGuardada = tareaGuardada[indicadorTarea]["clasesTarea"];
-		const keysClasesTareaGuardada = Object.keys(listaClasesTareaGuardada); // esto es un objeto
+		const keysClasesTareaGuardada = Object.keys(listaClasesTareaGuardada);
 		const cantidadClasesTareaGuardada = keysClasesTareaGuardada.length;
 		const nombreTareaGuardada = tareaGuardada[indicadorTarea]["nombreTarea"];
 		let listaALaQuePerteneceLaTareaGuardada;
@@ -368,101 +113,4 @@ function gestionarDataTareasGuardadas(dataTareas){
 
 		recrearTareasGuardadas(listaALaQuePerteneceLaTareaGuardada, clasesTareaGuardada, nombreTareaGuardada);
 	});
-}
-
-function recrearTareasGuardadas(listaALaQuePerteneceLaTareaGuardada, clasesTareaGuardada, nombreTareaGuardada){
-	const $listaTareas = document.querySelector(`.${listaALaQuePerteneceLaTareaGuardada}`);
-
-	const nuevaTarea = document.createElement("li");
-	nuevaTarea.classList = clasesTareaGuardada;
-
-	const nuevoContenedorEstado = document.createElement("div");
-	nuevoContenedorEstado.className = "contenedor-estado-tarea";
-
-	const nuevoEstadoTarea = document.createElement("input");
-	nuevoEstadoTarea.className = "form-check-input estado-tarea";
-	nuevoEstadoTarea.type = "checkbox";
-	nuevoEstadoTarea.id = "flexCheckDefault";
-
-	if(listaALaQuePerteneceLaTareaGuardada === "lista-tareas-completas") {
-		nuevoEstadoTarea.checked = true;
-	}
-
-	const nuevoTituloTarea = document.createElement("p");
-	nuevoTituloTarea.className = "nombre-tarea";
-	nuevoTituloTarea.textContent = nombreTareaGuardada;
-
-	const nuevoContenedorOpciones = document.createElement("div");
-	nuevoContenedorOpciones.className = "contenedor-opciones-tarea";
-
-	const nuevoBotonOpciones = document.createElement("button");
-	nuevoBotonOpciones.className = "boton-opciones-tarea";
-	nuevoBotonOpciones.type="button dropdown-toggle";
-	nuevoBotonOpciones.setAttribute("data-bs-toggle", "dropdown");
-
-	const nuevoIconoBoton = document.createElement("i");
-	nuevoIconoBoton.className = "fa-solid fa-ellipsis icono-opciones-tareas";
-
-	const nuevaListaOpciones = document.createElement("ul");
-	nuevaListaOpciones.className = "dropdown-menu";
-
-	const nuevaOpcionPrioridadTres = document.createElement("li");
-	nuevaOpcionPrioridadTres.className = "bloque-opciones-tareas";
-
-	const nuevaPrioridadTres = document.createElement("div");
-	nuevaPrioridadTres.className = "opciones-tareas opcion-prioridad-3";
-	nuevaPrioridadTres.textContent = "Prioridad: 3";
-
-	const nuevaOpcionPrioridadDos = document.createElement("li");
-	nuevaOpcionPrioridadDos.className = "bloque-opciones-tareas";
-
-	const nuevaPrioridadDos = document.createElement("div");
-	nuevaPrioridadDos.className = "opciones-tareas opcion-prioridad-2";
-	nuevaPrioridadDos.textContent = "Prioridad: 2";
-
-	const nuevaOpcionPrioridadUno = document.createElement("li");
-	nuevaOpcionPrioridadUno.className = "bloque-opciones-tareas";
-
-	const nuevaPrioridadUno = document.createElement("div");
-	nuevaPrioridadUno.className = "opciones-tareas opcion-prioridad-1";
-	nuevaPrioridadUno.textContent = "Prioridad: 1";
-
-	const nuevoBloqueEliminarTarea = document.createElement("li");
-	nuevoBloqueEliminarTarea.className = "bloque-opciones-tareas";
-
-	const nuevoContenedorBorrarTarea = document.createElement("div");
-	nuevoContenedorBorrarTarea.className = "bloque-borrar-tarea";
-	
-	const nuevoIconoBorrarTarea = document.createElement("i");
-	nuevoIconoBorrarTarea.className = "fa-regular fa-trash-can borrar-tarea";
-
-	const nuevoTextoBorrarTarea = document.createElement("div");
-	nuevoTextoBorrarTarea.className = "opciones-tareas borrar-tarea";
-	nuevoTextoBorrarTarea.textContent = "Eliminar tarea";
-
-	nuevoContenedorEstado.appendChild(nuevoEstadoTarea);
-	nuevoContenedorEstado.appendChild(nuevoTituloTarea);
-	nuevaTarea.appendChild(nuevoContenedorEstado);
-
-	nuevoBotonOpciones.appendChild(nuevoIconoBoton);
-	nuevaOpcionPrioridadTres.appendChild(nuevaPrioridadTres);
-	nuevaOpcionPrioridadDos.appendChild(nuevaPrioridadDos);
-	nuevaOpcionPrioridadUno.appendChild(nuevaPrioridadUno);
-	nuevoContenedorBorrarTarea.appendChild(nuevoIconoBorrarTarea);
-	nuevoContenedorBorrarTarea.appendChild(nuevoTextoBorrarTarea);
-	nuevoBloqueEliminarTarea.appendChild(nuevoContenedorBorrarTarea);
-	
-	nuevaListaOpciones.appendChild(nuevaOpcionPrioridadTres);
-	nuevaListaOpciones.appendChild(nuevaOpcionPrioridadDos);
-	nuevaListaOpciones.appendChild(nuevaOpcionPrioridadUno);
-	nuevaListaOpciones.appendChild(nuevoBloqueEliminarTarea);
-
-	nuevoContenedorOpciones.appendChild(nuevoBotonOpciones);
-	nuevoContenedorOpciones.appendChild(nuevaListaOpciones);
-	nuevaTarea.appendChild(nuevoContenedorOpciones);
-
-	$listaTareas.appendChild(nuevaTarea);
-
-	cambiarColorNombreTareas(nuevaTarea);
-	cambiarColorIconoOpcionesTareas(nuevaTarea);
 }
